requirements.txt
streamlit
llama-index
llama-index-core
llama-index-embeddings-google-genai
llama-index-llms-google-genai
llama-index-vector-stores-chroma
chromadb
python-dotenv
google-generativeai
